Split integration tests and clear Clippy warnings
Move shared HTTP test helpers into tests/support/mod.rs and split the
~1900-line integration.rs into focused binaries (health, rooms, UI,
RPC, HTML). Apply cargo clippy --fix plus manual cleanups so the
workspace builds with zero Clippy warnings.
