+++ server/src/html/forum/profile.rs
@@ -110,7 +110,7 @@ pub async fn user_profile_page(
                     }
                 }
             }
-            (cli_panel(&[format!("npx slugsocial public forum list")]))
+            (cli_panel(&["npx slugsocial public forum list".to_string()]))
         },
         Some(view_count),
         theme_from_jar(&jar),
