+++ server/src/html/garden.rs
@@ -233,6 +233,7 @@ fn vote_edge_history_markup(content: &ContentState, left: &ItemId, right: &ItemI
 }
 
 /// After a successful vote post: refresh edge history (no in-page preview card).
+#[allow(clippy::too_many_arguments)]
 pub(crate) async fn vote_compare_post_success_js(
     state: &AppState,
     nav: &ThreadNav,
@@ -1493,8 +1494,8 @@ async fn render_scope_view(
                 }
             }
             @let cli = match &scope {
-                ScopeId::Public => format!("npx slugsocial public garden body {}", cli_path_arg),
-                ScopeId::Room(room_id) => format!("npx slugsocial private {room_id} garden body {}", cli_path_arg),
+                ScopeId::Public => format!("npx slugsocial public garden body {cli_path_arg}"),
+                ScopeId::Room(room_id) => format!("npx slugsocial private {room_id} garden body {cli_path_arg}"),
             };
             (cli_panel(std::slice::from_ref(&cli)))
         },
